로깅 및 트랜잭션 개선, 프로파일 설정 - 서비스, 컨트롤러 클래스에 `@Slf4j` 추가 및 주요 로직에 로그 추가 (INFO/DEBUG/WARN/ERROR 레벨 활용) - 기존 클래스 레벨에 트랜잭션(@Transactional) 선언하던 방식을 메서드 레벨에 선언하도록 변경 - 개발(dev), 운영(prod) 환경에 맞춘 로깅 레벨 설정 추가 (콘솔 및 파일 롤링 정책 설정) - Logback 설정 파일 구성(로깅 패턴, 출력 방식 커스터마이징, 콘솔과 파일에 동시에 로그 기록, 일자별 로그 파일 롤링, 최대 30일간 보관 설정) - 파일 다운로드 API에 Swagger 문서화 수정 - 페이지네이션 처리 방식 개선 (기본값 추가 및 Pageable 인자 활용) - sql문 불필요한 , 오타 제거

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -15,6 +15,7 @@ import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
@@ -23,6 +24,7 @@ import java.util.NoSuchElementException;
 import java.util.UUID;
 import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @RequiredArgsConstructor // 의존성 주입
 @Service // Basic*Service 구현체를 Service 인터페이스의 Bean으로 등록
 @Transactional(readOnly = true)
@@ -42,9 +44,11 @@ public class BasicUserService implements UserService {
     String username = userCreateRequest.getUsername();
     String email = userCreateRequest.getEmail();
     if (userRepository.existsByEmail(email)) { // API 스펙에 맞춰 추가
+      log.warn("이미 사용 중인 사용자 이메일 {}", email);
       throw new IllegalArgumentException("User with email " + email + " already exists");
     }
     if (userRepository.existsByUsername(username)) {
+      log.warn("이미 사용 중인 사용자 이름 {}", username);
       throw new IllegalArgumentException("User with username " + username + " already exists");
     }
     // 프로필 이미지 있으면 등록
@@ -60,22 +64,26 @@ public class BasicUserService implements UserService {
         })
         .orElse(null);
     // 프로필 이미지 없으면 이는 비워두고 등록
-    User user = new User(userCreateRequest.getUsername(), userCreateRequest.getEmail(),
+    User user = new User(username, email, // 중복된 호출 변수 사용
         userCreateRequest.getPassword(), savedProfile);
-    User savedUser = userRepository.save(user);
     // UserStatus를 같이 생성
-    UserStatus userStatus = new UserStatus(savedUser, Instant.now());
-    statusRepository.save(userStatus);
+    UserStatus userStatus = new UserStatus(user, Instant.now());
+    User savedUser = userRepository.save(user);
+    // 영속성 전이-> user만 저장해도 되는 것
+    log.info("사용자 등록 userName={}, userEmail={}, savedProfile={}", username, email, savedProfile);
     return userMapper.toDto(savedUser);
   }
 
+  @Transactional(readOnly = true)
   @Override
   public UserDto find(UUID userId) {
-    User user = userRepository.findById(userId)
+    log.debug("사용자 조회 userId={}", userId);
+    return userRepository.findById(userId)
+        .map(user -> userMapper.toDto(user))
         .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
-    return userMapper.toDto(user);
   }
 
+  @Transactional(readOnly = true)
   @Override
   public List<UserDto> findAll() {
     return userRepository.findAll().stream()
@@ -92,16 +100,16 @@ public class BasicUserService implements UserService {
     String newEmail = userUpdateRequest.getNewEmail();
     if (userRepository.existsByEmail(newEmail) && !user.getEmail()
         .equals(newEmail)) { // API 스펙에 맞춰 추가
+      log.warn("이미 사용 중인 사용자 이메일 {}", newEmail);
       throw new IllegalArgumentException("User with email " + newEmail + " already exists");
     }
     if (userRepository.existsByUsername(newUsername)) {
+      log.warn("이미 존재하는 사용자 이름 {}", newUsername);
       throw new IllegalArgumentException("User with username " + newUsername + " already exists");
     }
     // 프로필 이미지 선택적으로 대체
     BinaryContent profile = optionalProfileCreateRequest
         .map(profileRequest -> {
-          Optional.ofNullable(user.getProfile()).ifPresent(contentRepository::delete);
-
           BinaryContent binaryContent = new BinaryContent(
               profileRequest.getFileName(),
               profileRequest.getContentType(),
@@ -112,20 +120,21 @@ public class BasicUserService implements UserService {
         })
         .orElse(null);
     // 기존 프로필 삭제
-    String newPassword = userUpdateRequest.getNewPassword();
-    user.update(newUsername, newEmail, newPassword, profile); // profileId 추가
+    user.update(newUsername, newEmail, userUpdateRequest.getNewPassword(), profile); // profileId 추가
+    log.info("사용자 정보 수정 userId={}, newUserName={}, newEmail, fileName={}", userId,
+        newUsername, newEmail, profile.getFileName());
     return userMapper.toDto(user);
   }
 
   @Transactional
   @Override
   public void delete(UUID userId) {
-    User user = userRepository.findById(userId)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
-    // 관련된 도메인도 같이 삭제
-    if (user.getProfile() != null) {
-      contentRepository.delete(user.getProfile());
+    if (userRepository.existsById(userId)) {
+      log.warn("존재하지 않는 사용자 아이디 {}", userId);
+      throw new NoSuchElementException("User with id " + userId + " not found");
     }
+    // user가 삭제되면 영속성 전이된 프로필 또한 삭제
     userRepository.deleteById(userId);
+    log.info("사용자 삭제 userId={}", userId);
   }
 }
